requirements.txt
streamlit
ultralytics
Pillow
